Refactored dataloader tests
